refactor: 마지막 페이지 확인 함수 utils로 이동
